iot/requirements.txt
azure-iot-device==2.12.0
geopy==2.3.0
paho-mqtt==1.6.1